pyproject.toml
[project]
name = "rule-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
